# docs/api/openapi.yml
openapi: 3.0.0
info:
  title: Defcon API
  description: Defcon control plane API documentation
  version: 0.1.0
...
paths:
...
  /api/checks/{uuid}/outages:
    get:
      summary: Get a check outages
      tags: ['Checks']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
        - name: from
          description: Display outages from this date
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          description: Display outages until this date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of outages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/outage/Outage'
  /api/checks/{uuid}/events:
    get:
      summary: Get a check outage events
      tags: ['Checks']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
        - name: from
          description: Display events from this date
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          description: Display events until this date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of handler events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/site_outage/Event'
...
  /api/outages:
    get:
      summary: List current outages
      tags: ['Outages']
      parameters:
        - name: from
          description: Display resolved outages from this date
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          description: Display resolved outages until this date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of outages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/outage/Outage'
  /api/outages/{uuid}/events:
    get:
      summary: Get a site outage events
      tags: ['Site outages']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of handler events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/site_outage/Event'
  /api/outages/{uuid}/comment:
    put:
      summary: Add a comment
      tags: ['Outages']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        description: The comment to be added to the outage
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  description: The comment to be added
                  type: string
      responses:
        '200':
          description: The comment was recorded
        '400':
          description: Invalid payload
        '404':
          description: Provided UUID does not match a known alerter
...